Correções.

+++ Gestor-de-Modulos/Gestor-de-Modulos/7_gdm_modulos_capitalizados.sql
@@ -6,7 +6,7 @@ CREATE TABLE modulos_capitalizados(
     id INT(11) UNSIGNED NOT NULL AUTO_INCREMENT,
     id_curso_disciplina_modulo INT(11) UNSIGNED NOT NULL,
     id_aluno INT(11) UNSIGNED NOT NULL,
-    nota INT UNSIGNED NOT NULL,
+    nota TINYINT(3) UNSIGNED NOT NULL,
     dt_capatilazado DATE NOT NULL,
     PRIMARY KEY (id),
     CONSTRAINT fk_modulos_capitalizados_alunos
